refactor: :recycle: Refactor to use a single animated ellipsis component
