+++ gui/src/components/mainInput/Lump/LumpToolbar/GeneratingIndicator.tsx
@@ -1,4 +1,4 @@
-import { AnimatedEllipsis } from "../../..";
+import { AnimatedEllipsis } from "../../../AnimatedEllipsis";
 
 export function GeneratingIndicator({
   text = "Generating",
